MonetaryTreasure
I put in some text which if correct (is it?) tells me some information about things. In this one I put in info about declaring a class and what goes in the class.
Also expanded on the summary of the properties to try to make them more precise and clear.
Also put in some todo's with questions.